feat: componentization and add mobile tab
