+++ src/common/lib/theme/horizontalTabs.css
@@ -1,0 +1,9 @@
+/* Esconde a barra de rolagem horizontal para vários navegadores */
+.scrollbar-hide::-webkit-scrollbar {
+  display: none;
+}
+
+.scrollbar-hide {
+  -ms-overflow-style: none;  /* IE e Edge */
+  scrollbar-width: none;  /* Firefox */
+}
